requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
optuna>=3.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
scipy>=1.9.0
category-encoders>=2.5.0
gradio>=3.40.0
pyyaml>=6.0
joblib>=1.2.0
gdown>=4.7.0
pytest>=7.2.0